requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
tensorflow==2.15.1
mediapipe==0.10.14
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
scikit-learn==1.3.2
Werkzeug==3.0.1
gunicorn==21.2.0
